CryptoCurrency methods are injected to Currency methods. Unnecessary classes are deleted.
