+++ backend/src/main/java/com/eyetrade/backend/constants/CurrencyType.java
@@ -6,5 +6,11 @@ public enum CurrencyType {
     EUR,
     GBP,
     JPY,
-    CNY
+    CNY,
+    BTC,
+    ETH,
+    XRP,
+    LTC,
+    XMR
+
 }
