--- backend/src/main/java/com/eyetrade/backend/model/dto/currency/CryptoCurrencyConverterDto.java
@@ -1,17 +1,0 @@
-package com.eyetrade.backend.model.dto.currency;
-
-import com.eyetrade.backend.constants.CryptoCurrencyType;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class CryptoCurrencyConverterDto {
-
-    private CryptoCurrencyType source;
-    private CryptoCurrencyType target;
-    private Double amount;
-
-}
