--- backend/src/main/java/com/eyetrade/backend/model/resource/currency/CryptoCurrencyIntervalResource.java
@@ -1,18 +1,0 @@
-package com.eyetrade.backend.model.resource.currency;
-
-import lombok.*;
-
-import javax.annotation.Resource;
-import java.util.List;
-
-@Resource
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class CryptoCurrencyIntervalResource {
-
-    List<CryptoCurrencyConverterResource> resources;
-    String startDate;
-    String endDate;
-
-}
