+++ backend/src/main/java/com/eyetrade/backend/service/CurrencyRecordService.java
@@ -1,5 +1,6 @@
 package com.eyetrade.backend.service;
 
+import com.eyetrade.backend.constants.CryptoCurrencyType;
 import com.eyetrade.backend.constants.CurrencyConstants;
 import com.eyetrade.backend.constants.CurrencyType;
 import com.eyetrade.backend.model.entity.CurrencyRecord;
@@ -9,6 +10,7 @@ import org.apache.http.HttpEntity;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.methods.HttpUriRequest;
+import org.apache.http.client.utils.URIBuilder;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,6 +20,8 @@ import org.springframework.stereotype.Service;
 import javax.transaction.Transactional;
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
@@ -53,6 +57,17 @@ public class CurrencyRecordService {
         record.setSterlingRate(rates.get(CurrencyType.GBP.toString()).doubleValue());
         record.setJapanRate(rates.get(CurrencyType.JPY.toString()).doubleValue());
         record.setChinaRate(rates.get(CurrencyType.CNY.toString()).doubleValue());
+
+        Double dollarRate=record.getDollarRate();
+        try {
+            record.setBitcoin(getCryptoCurrenciesExchangeRate(CryptoCurrencyType.BTC)/dollarRate);
+            record.setEthereum(getCryptoCurrenciesExchangeRate(CryptoCurrencyType.ETH)/dollarRate);
+            record.setLitecoin(getCryptoCurrenciesExchangeRate(CryptoCurrencyType.LTC)/dollarRate);
+            record.setMonero(getCryptoCurrenciesExchangeRate(CryptoCurrencyType.XMR)/dollarRate);
+            record.setRipple(getCryptoCurrenciesExchangeRate(CryptoCurrencyType.XRP)/dollarRate);
+        } catch (URISyntaxException e) {
+            e.printStackTrace();
+        }
         currencyRepository.save(record);
         transactionService.checkBuySellOrder();
         return record;
@@ -92,6 +107,40 @@ public class CurrencyRecordService {
 
     private boolean checkExpired(Long lastUpdateTime){
         return (new Date().getTime()-lastUpdateTime>CURRENCY_EXPIRE_TIME);
+    }
+
+    private Double getCryptoCurrenciesExchangeRate(CryptoCurrencyType toCurrency) throws IOException, URISyntaxException {
+        // TODO: We should decide what to do with that unstable api
+        // give 4 chance to the api
+        for(int i = 0; i < 3; i++){
+            try{
+                return getExchangeRateFromApi(toCurrency);
+            }
+            catch (Exception e) {
+            }
+        }
+        return getExchangeRateFromApi(toCurrency);
+    }
+
+    private Double getExchangeRateFromApi(CryptoCurrencyType toCurrency) throws URISyntaxException, IOException {
+        URI apiUri = new URIBuilder()
+                .setScheme("https")
+                .setHost(CurrencyConstants.CRYPTO_CURRENCY_API_HOST)
+                .setPath(CurrencyConstants.CRYPTO_CURRENCY_API_PATH)
+                .setParameter("function", CurrencyConstants.CRYPTO_CURRENCY_API_EXCHANGE_FUNCTION)
+                .setParameter("from_currency", CurrencyConstants.CRYPTO_CURRENCY_API_BASE)
+                .setParameter("to_currency", toCurrency.toString())
+                .setParameter("apikey", CurrencyConstants.CRYPTO_CURRENCY_API_KEY)
+                .build();
+
+        HttpUriRequest request = new HttpGet(apiUri);
+        try (CloseableHttpResponse response = httpClient.execute(request)) {
+            HttpEntity entity = response.getEntity();
+            InputStream content = entity.getContent();
+            Map<?, ?> data = objectMapper.readValue(content, LinkedHashMap.class);
+            String rateInStr = ((Map<String, String>) data.get("Realtime Currency Exchange Rate")).get("5. Exchange Rate");
+            return Double.parseDouble(rateInStr);
+        }
     }
